feat(bt): add tentative rough construction
