+++ succinct_neo/Cargo.toml
@@ -22,6 +22,7 @@ paste = "1.0.12"
 rand = { version = "0.8.5", features = [ "small_rng" ]}
 gcd = "2.3"
 id-arena = "2.2.1"
+multimap = "0.9.0"
 
 [dev-dependencies]
 criterion = "0.4.0"
