+++ succinct_neo/src/lib.rs
@@ -5,5 +5,5 @@
 pub mod bit_vec;
 pub mod int_vec;
 pub mod rank_select;
-pub mod traits;
 pub mod rolling_hash;
+pub mod traits;
